add member, practice components, router basics
