+++ src/app/app.component.scss
@@ -1,0 +1,17 @@
+.toolbar-tools {
+  display: flex;
+  width: 100%;
+}
+
+.mat-icon {
+  height: 19px;
+  padding-right: 10px;
+}
+
+#title {
+  display: inline-block;
+  padding-left: 10px;
+  line-height: 48px;
+  font-size: 24px;
+  font-weight: 500;
+}
