+++ src/app/models/practice.ts
@@ -1,0 +1,19 @@
+export interface IPractice {
+  id: string;
+  practiceNumber: number;
+  practiceDate: Date;
+  practiceTopic: string;
+  practiceCost: number;
+  miscExpense: number;
+  miscRevenue: number;
+}
+
+export class Practice implements IPractice {
+  id: string;
+  practiceNumber: number;
+  practiceDate: Date;
+  practiceTopic: string;
+  practiceCost: number;
+  miscExpense: number;
+  miscRevenue: number;
+}
