+++ src/app/models/punch-card.ts
@@ -1,0 +1,13 @@
+export interface IPunchCard {
+  id: string;
+  purchaseMemberId: string;
+  purchaseDate: Date;
+  purchaseAmount: number;
+}
+
+export class PunchCard implements IPunchCard {
+  id: string;
+  purchaseMemberId: string;
+  purchaseDate: Date;
+  purchaseAmount: number;
+}
