+++ src/app/models/punch.ts
@@ -1,0 +1,9 @@
+export interface IPunch {
+  id: string;
+  attendanceId: string;
+}
+
+export class Punch implements IPunch {
+  id: string;
+  attendanceId: string;
+}
