+++ src/app/practices/practices/practices.component.html
@@ -1,0 +1,1 @@
+<p>practices works!</p>
