+++ src/app/practices/practices/practices.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-practices',
+  templateUrl: './practices.component.html',
+  styleUrls: ['./practices.component.scss']
+})
+export class PracticesComponent {
+
+}
